pyproject.toml
[tool.poetry]
name = "canopy-sdk"
version = "0.5.0"
description = "Retrieval Augmented Generation (RAG) framework and context engine powered by Pinecone"
authors = ["Relevance Team <relevance@pinecone.io>"]
readme = "README.md"
license = "Apache-2.0"
packages = [{include = "canopy", from = "src"},
            {include = "canopy_cli", from = "src"},
            {include = "canopy_server", from = "src"},]

[tool.poetry.dependencies]
python = "^3.9"
pinecone-client = "^2.2.2"
python-dotenv = "^1.0.0"
openai = "^1.2.3"
tiktoken = "^0.3.3"
pinecone-datasets = "^0.6.2"
pydantic = "^1.10.7"
pandas-stubs = "^2.0.3.230814"
fastapi = ">=0.92.0, <1.0.0"
uvicorn = ">=0.20.0, <1.0.0"
tenacity = "^8.2.1"
sse-starlette = "^1.6.5"
types-tqdm = "^4.61.0"
tqdm = "^4.66.1"
gunicorn = "^21.2.0"
types-pyyaml = "^6.0.12.12"
jsonschema = "^4.2.0"
types-jsonschema = "^4.2.0"
prompt-toolkit = "^3.0.39"
pinecone-text = "^0.7.1"
tokenizers = "^0.15.0"
transformers = "^4.35.2"
sentencepiece = "^0.1.99"
cohere = { version = ">=4.37", optional = true }
pyjwt = "^2.8.0"

[tool.poetry.extras]
cohere = ["cohere"]

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pytest = "^7.3.2"
mypy = "^1.4.1"
flake8 = "^6.1.0"
pytest-html = "^4.1.0"
flake8-pyproject = "^1.2.3"
asyncio = "^3.4.3"
pytest-asyncio = "^0.14.0"
pytest-mock = "^3.6.1"
pytest-xdist = "^3.3.1"
types-requests = "^2.31.0.2"
httpx = "^0.25.0"
pydoclint = "^0.3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
allow_redefinition = true
exclude = ['tests', '.venv']

[[tool.mypy.overrides]]
module = [
    'pinecone_text.*',
    'pinecone_datasets',
    'pinecone',
    'transformers.*'
]
ignore_missing_imports = true


[tool.flake8]
per-file-ignores = [
    '*/__init__.py:F401',
]
exclude = ['.venv']
max-line-length = 88

# PyDocLint configuration
style = 'google'
arg-type-hints-in-docstring = false
require-return-section-when-returning-nothing = false
allow-init-docstring = true
check-return-types = false
skip-checking-raises = true

[tool.poetry.scripts]
canopy = "canopy_cli.cli:cli"

[tool.pytest.ini_options]
log_cli = true